refactor(runtime): split exec vm helpers
